Refactor

+++ core/src/main/java/io/github/game/ui/screens/LoadingScreen.java
@@ -45,8 +45,8 @@ public class LoadingScreen implements Screen {
         // Отрисовка прогресса загрузки
         spriteBatch.begin();
         font.draw(spriteBatch, "Loading: " + (int) (progress * 100) + "%",
-                  Gdx.graphics.getWidth() / 2 - 50,
-                  Gdx.graphics.getHeight() / 2);
+                  (float) Gdx.graphics.getWidth() / 2 - 50,
+                  (float) Gdx.graphics.getHeight() / 2);
         spriteBatch.end();
 
         // Если загрузка завершена, переключаемся на главный экран
